Time: 5 ms (83.58%), Space: 46 MB (33.35%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.java
@@ -5,7 +5,7 @@ class Solution {
         }
 
         int og = x;
-        int rev = 0;
+        long rev = 0;
 
         while (x > 0) {
             int digit = x % 10;
